// cloudflare/src/types/api.ts
export interface LatitudeLongitude {
  latitude: number;
  longitude: number;
}

export type MeteredTimeResolution = "MINUTES" | "FIFTEEN_MINUTES" | "HOURS" | "DAYS" | "WEEKS" | "MONTHS" | "YEARS";

export interface LocationInformation {
  heatSource?: string;
  locationSqFt?: string;
  numElectricCars?: string;
  locationType?: string;
  hotTub?: string;
  airConditioning?: string;
  swimmingPool?: string;
  numPeople?: string;
  primaryVehicleBrand?: string | null;
  [key: string]: any;
}
// ...
export interface Outlet {
  deviceGid: number;
  outletOn: boolean;
  loadGid: number;
}

export interface EvCharger {
  deviceGid: number;
  loadGid: number;
  message: string;
  status: string;
  icon: string;
  iconLabel: string;
  debugCode: string;
  iconDetailText: string;
  faultText: string | null;
  proControlCode: string | null;
  breakerPIN: string | null;
  chargerOn: boolean;
  chargingRate: number;
  maxChargingRate: number;
  loadManagementEnabled: boolean;
  hideChargeRateSliderText: string | null;
}

export interface Battery {
  deviceGid: number;
  loadGid: number;
  capacityKWh: number;
  inverterMaxPowerKW: number;
  maxActivePowerKW: number;
  emsSN: string;
  emsVersion: string;
  inverterSN: string;
  inverterVersion: string;
  status: string;
  solarIcon: string;
  batteryIcon: string;
  gridIcon: string;
  powerFlow: any[];
  allowExport: boolean;
  hasSolar: boolean;
  onlyChargeFromExcessSolar: boolean;
  reserveStateOfCharge: number;
  stateOfCharge: number;
  batteryPowerKW: number;
  chargePowerKW: number;
  dischargePowerKW: number;
  inverterPowerKW: number;
  gridPowerKW: number;
  powerFromGridKW: number;
  powerToGridKW: number;
  pvPowerKW: number;
  loadPowerKW: number;
  peakStatus: string;
  peakRateKW: number;
  offPeakStatus: string;
  offPeakRateKW: number;
  offPeakEndChargeDuration: string;
  offPeakEndChargeRateKW: number;
  batteryControllable: boolean;
  maxGridPowerKW: number;
}
// ...
export interface LocationSummary {
  latitude_longitude: LatitudeLongitude | null;
  time_zone: string;
  location_information: LocationInformation | null;
  zip_code: string | null;
}

export interface DeviceInfo {
  device_gid: number;
  manufacturer_device_id: string;
  model: string;
  firmware_version: string | null;
  connected: boolean;
  offline_since: string | null;
  device_type: string;
  display_name: string;
  utility_rate_selected: boolean;
  usage_cents_per_kwh: number | null;
  billing_cycle_start_day: number;
  location_summary: LocationSummary;
  device_details: {
    ev_charger: EvCharger | null;
    smart_plug: Outlet | null;
    battery: Battery | null;
  };
}